systemui: never show excluded recent task in recents view
Actually, when there there isn't recent tasks to show, the systemui forces to show
task with the flag FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS, but only the first one.
This create an non-consistent behaviour when the user remove all recent task and
then there is a new recent task in the recents view, kill that and a new one
appears, an so on, until no task are available.

This patch changes that behaviour to:
- Never show tasks in recents view with the flag FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS
- Clear recents button kill all the tasks (included the ones with FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS)

Change-Id: I8dc2859ca9dcde7229d5fdb502da313b1ae0dccb
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/recents/views/TaskStackView.java

Signed-off-by: Chet Kener <Cl3Kener@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/recents/views/TaskStackView.java